added fb_write to a header file
